Syncbranch

+++ src/LuaEngine/methods/PlayerMethods.h
@@ -531,6 +531,13 @@ namespace LuaPlayer
         return 1;
     }
 
+    /**
+     * Returns `true` if the [Player] is in the same group and visible to the specified [Player], `false` otherwise.
+     *
+     * @param [Player] player : the source player
+     * @param [Player] target : the player to check visibility from
+     * @return bool isGroupVisible
+     */
     int IsGroupVisibleFor(lua_State* L, Player* player)
     {
         Player* target = Eluna::CHECKOBJ<Player>(L, 2);
@@ -592,6 +599,12 @@ namespace LuaPlayer
         return 1;
     }
 
+    /**
+     * Returns `true` if the [Player] is currently visible to other players, `false` if hidden via GM invisibility.
+     *
+     * @param [Player] player
+     * @return bool isVisible
+     */
     int IsGMVisible(lua_State* L, Player* player)
     {
         Eluna::Push(L, player->isGMVisible());
@@ -609,6 +622,12 @@ namespace LuaPlayer
         return 1;
     }
 
+    /**
+     * Returns `true` if the [Player] has GM chat enabled, `false` otherwise.
+     *
+     * @param [Player] player
+     * @return bool isGMChat
+     */
     int IsGMChat(lua_State* L, Player* player)
     {
         Eluna::Push(L, player->isGMChat());
@@ -1683,7 +1702,7 @@ namespace LuaPlayer
             {
                 if (mask & (1 << bit))
                 {
-                    uint8 nodeId = (i * 32) + bit + 1;
+                    uint32 nodeId = (i * 32) + bit + 1;
                     lua_pushinteger(L, nodeId);
                     lua_rawseti(L, -2, lua_rawlen(L, -2) + 1);
                 }
@@ -2480,7 +2499,7 @@ namespace LuaPlayer
      *
      * @param bool takeCost = true
      * @param float discountMod = 1.0
-     * @param bool guidBank = false
+     * @param bool guildBank = false
      */
     int DurabilityRepairAll(lua_State* L, Player* player)
     {
@@ -3192,6 +3211,12 @@ namespace LuaPlayer
         return 1;
     }
 
+    /**
+     * Adds a specified number of lifetime honorable kills to the [Player].
+     *
+     * @param [Player] player
+     * @param uint32 kills
+     */
     int AddLifetimeKills(lua_State* L, Player* player)
     {
         uint32 val = Eluna::CHECKVAL<uint32>(L, 2);
@@ -3486,9 +3511,10 @@ namespace LuaPlayer
     }
 
     /**
-    * Get glyphId of the glyph slot specified by `slotIndex` off the [Player]'s current talent specialization.`
-    * @param uint32 slotIndex
-    * @return glyphId of the glyph in the selected glyph slot or 0 in case the glyph slot is empty
+    * Returns the glyph ID in the specified glyph slot of the [Player]'s current talent specialization.
+    *
+    * @param [uint32] slotIndex
+    * @return [uint32] glyphId
     */
     int GetGlyph(lua_State* L, Player* player)
     {
@@ -4098,4 +4124,3 @@ namespace LuaPlayer
     }
 };
 #endif
-
